requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.26.4
azure-cosmos==4.3.1
azure-identity==1.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msal==1.22.0
msal-extensions==1.0.0
multidict==6.0.4
numpy==1.24.3
openai==0.27.4
pandas==2.0.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
regex==2023.3.23
requests==2.28.2
six==1.16.0
tiktoken==0.3.3
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
Win32Security==2.1.0
yarl==1.9.1
zipp==3.15.0
